reorganized frontend file structure
Created a new folder 'pages' to hold files that are rendered as pages. Moved login page logic from App.js --> new file LoginPage.js and renamed Dashboard.js --> DashboardPage.js. Moved both these files into 'pages' folder.
